requirements/dev.txt
-r base.txt

pytest==5.4.3
pytest-django==3.9.0